// enjin/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  "agents/clipper": typeof agents_clipper;
  clips: typeof clips;
  files: typeof files;
  index: typeof index;
  steps: typeof steps;
  types: typeof types;
  "zod_schemas/clipper": typeof zod_schemas_clipper;
}>;
// ...
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;